Add validation rules decoding.
