+++ src/xcode/ENA/HealthCertificateToolkit/Package.swift
@@ -5,6 +5,9 @@ import PackageDescription
 
 let package = Package(
     name: "HealthCertificateToolkit",
+    platforms: [
+      .iOS(.v11)
+    ],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
@@ -16,14 +19,15 @@ let package = Package(
         .package(url: "https://github.com/unrelentingtech/SwiftCBOR", .upToNextMajor(from: "0.4.3")),
         .package(url: "https://github.com/corona-warn-app/base45-swift", .branch("distribution/swiftpackage")),
         .package(name: "JSONSchema", url: "https://github.com/eu-digital-green-certificates/JSONSchema.swift", .upToNextMajor(from: "0.6.0")),
-        .package(url: "https://github.com/tsolomko/SWCompression.git", .upToNextMajor(from: "4.5.0"))
+        .package(url: "https://github.com/tsolomko/SWCompression.git", .upToNextMajor(from: "4.5.0")),
+        .package(name: "CertLogic", url: "https://github.com/eu-digital-green-certificates/dgc-certlogic-ios", .branch("develop"))
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
             name: "HealthCertificateToolkit",
-            dependencies: ["SwiftCBOR", "base45-swift", "JSONSchema", "SWCompression"],
+            dependencies: ["SwiftCBOR", "base45-swift", "JSONSchema", "SWCompression", "CertLogic"],
             resources: [.process("Access/Ressources/CertificateSchema.json")]
         ),
         .testTarget(
